Wording and code organization fixes.
